Добавил папку App в которую перенес всю инициализирующую логику
