+++ frontend_fungi/index.html
@@ -8,6 +8,6 @@
   </head>
   <body>
     <div id="root"></div>
-    <script type="module" src="/src/main.tsx"></script>
+    <script type="module" src="/src/app/entrypoint/index.tsx"></script>
   </body>
 </html>
