+++ frontend_fungi/src/app/layout/index.css
