+++ frontend_fungi/src/app/providers/types.ts
@@ -1,0 +1,5 @@
+import { ReactNode } from 'react';
+
+export interface IWithProviderProps {
+    children: ReactNode;
+}
