+++ frontend_fungi/src/app/providers/with-query-client.tsx
@@ -1,0 +1,10 @@
+import { IWithProviderProps } from "./types";
+import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
+
+const queryClient = new QueryClient();
+
+export const WithQueryClient: React.FC<IWithProviderProps> = ({ children }) => {
+  return (
+    <QueryClientProvider client={queryClient}>{children}</QueryClientProvider>
+  );
+};
