+++ frontend_fungi/src/app/providers/with-router.tsx
@@ -1,0 +1,12 @@
+import { BrowserRouter } from "react-router-dom";
+import { IWithProviderProps } from "./types";
+
+export const WithRouter: React.FC<IWithProviderProps> = ({ children }) => {
+  return (
+    <BrowserRouter
+      future={{ v7_startTransition: true, v7_relativeSplatPath: true }}
+    >
+      {children}
+    </BrowserRouter>
+  );
+};
