+++ frontend_fungi/src/app/styles/index.css
